# headless/headless-commerce/headless-commerce-admin-catalog-impl/rest-openapi.yaml
components:
...
    schemas:
...
        Sku:
            properties:
                cost:
                    example: 101
                    format: bigdecimal
                    minimum: 0
                    type: number
                depth:
                    example: 1.1
                    format: double
                    minimum: 0
                    type: number
                displayDate:
                    example: 2017-07-21
                    format: date
                    type: string
                expirationDate:
                    example: 2017-08-21
                    format: date
                    type: string
                externalReferenceCode:
                    example: AB-34098-789-N
                    type: string
                gtin:
                    example: "12341234"
                    type: string
                height:
                    example: 20.2
                    format: double
                    minimum: 0
                    type: number
                id:
                    example: 30130
                    format: int64
                    minimum: 0
                    readOnly: true
                    type: integer
                inventoryLevel:
                    example: 29
                    readOnly: true
                    type: integer
                manufacturerPartNumber:
                    example: "12341234"
                    type: string
                neverExpire:
                    example: true
                    type: boolean
                options:
                    additionalProperties:
                        type: string
                    example:
                        color: yellow
                        optionKey: optionValueKey
                        size: xs
                    type: object
                price:
                    example: 101
                    format: bigdecimal
                    minimum: 0
                    type: number
                productId:
                    example: 30129
                    format: int64
                    minimum: 0
                    readOnly: true
                    type: integer
                promoPrice:
                    example: 202
                    format: bigdecimal
                    minimum: 0
                    type: number
                published:
                    example: true
                    type: boolean
                purchasable:
                    example: true
                    type: boolean
                sku:
                    example: "12341234"
                    type: string
                weight:
                    example: 1.1
                    format: double
                    minimum: 0
                    type: number
                width:
                    example: 20.2
                    format: double
                    minimum: 0
                    type: number
            required:
                - sku
            type: object
...
    securitySchemes:
        BasicAuth:
            scheme: basic
            type: http
        BearerAuth:
            scheme: bearer
            type: http
        OAuth2:
            flows:
                authorizationCode:
                    authorizationUrl: /oauth/authorize
                    scopes:
                        HeadlessCommerceAdminCatalog.admin: Grants access to admin operations
                        HeadlessCommerceAdminCatalog.read: Grants read access
                        HeadlessCommerceAdminCatalog.write: Grants write access
                    tokenUrl: /oauth/token
            type: oauth2
info:
    contact:
        email: team-commerce@liferay.com
        name: Commerce Team
    description:         Liferay Commerce Admin Catalog API
    license:
        name: "Apache 2.0"
        url: "http://www.apache.org/licenses/LICENSE-2.0.html"
    title: Liferay Commerce Admin Catalog API
    version: v1.0
openapi: 3.0.0
paths:
...
    /products/by-externalReferenceCode/{externalReferenceCode}/skus/:
        get:
            operationId: getProductByExternalReferenceCodeSkusPage
            parameters:
                - in: path
                  explode: false
                  name: externalReferenceCode
                  required: true
                  schema:
                      description:
                          The resource external reference code
                      example: AB-34098-789-N
                      minLength: 1
                      type: string
                  style: simple
                - in: query
                  explode: true
                  name: page
                  required: false
                  schema:
                      type: integer
                  style: form
                - in: query
                  explode: true
                  name: pageSize
                  required: false
                  schema:
                      type: integer
                  style: form
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: '#/components/schemas/Sku'
                                type: array
                        application/xml:
                            schema:
                                items:
                                    $ref: '#/components/schemas/Sku'
                                type: array
                    description:
                        Successful operation
                401:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description:
                        Authentication information is missing or invalid
                404:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description:
                        The specified resource was not found
                500:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description:
                        Unexpected error
            security:
                - BasicAuth: []
                - BearerAuth: []
                - OAuth2:
                      - HeadlessCommerceAdminCatalog.read
            summary: Gets a list of Skus related to a Product.
            tags:
                - Sku
        post:
            operationId: postProductByExternalReferenceCodeSku
            parameters:
                - in: path
                  explode: false
                  name: externalReferenceCode
                  required: true
                  schema:
                      description:
                          The resource external reference code
                      example: AB-34098-789-N
                      minLength: 1
                      type: string
                  style: simple
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/Sku'
                    application/xml:
                        schema:
                            $ref: '#/components/schemas/Sku'
                required: true
            responses:
                201:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Sku'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/Sku'
                    description:
                        Created
                202:
                    description:
                        Accepted - Async
                204:
                    description:
                        Updated
                400:
                    description:
                        Invalid input
                401:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description:
                        Authentication information is missing or invalid
                404:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description:
                        The specified resource was not found
                500:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description:
                        Unexpected error
            security:
                - BasicAuth: []
                - BearerAuth: []
                - OAuth2:
                      - HeadlessCommerceAdminCatalog.write
            summary: Creates or updates a Sku related to a Product.
            tags:
                - Sku
...
    /products/{id}/skus/:
        get:
            operationId: getProductIdSkusPage
            parameters:
                - in: path
                  explode: false
                  name: id
                  required: true
                  schema:
                      description:
                          The resource Id
                      example: 30130
                      format: int64
                      minimum: 0
                      type: integer
                  style: simple
                - in: query
                  explode: true
                  name: page
                  required: false
                  schema:
                      type: integer
                  style: form
                - in: query
                  explode: true
                  name: pageSize
                  required: false
                  schema:
                      type: integer
                  style: form
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: '#/components/schemas/Sku'
                                type: array
                        application/xml:
                            schema:
                                items:
                                    $ref: '#/components/schemas/Sku'
                                type: array
                    description:
                        Successful operation
                401:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description:
                        Authentication information is missing or invalid
                404:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description:
                        The specified resource was not found
                500:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description:
                        Unexpected error
            security:
                - BasicAuth: []
                - BearerAuth: []
                - OAuth2:
                      - HeadlessCommerceAdminCatalog.read
            summary: Gets a list of Skus related to a Product.
            tags:
                - Sku
        post:
            operationId: postProductIdSku
            parameters:
                - in: path
                  explode: false
                  name: id
                  required: true
                  schema:
                      description:
                          The resource Id
                      example: 30130
                      format: int64
                      minimum: 0
                      type: integer
                  style: simple
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/Sku'
                    application/xml:
                        schema:
                            $ref: '#/components/schemas/Sku'
                required: true
            responses:
                201:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Sku'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/Sku'
                    description:
                        Created
                202:
                    description:
                        Accepted - Async
                204:
                    description:
                        Updated
                400:
                    description:
                        Invalid input
                401:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description:
                        Authentication information is missing or invalid
                404:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description:
                        The specified resource was not found
                500:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description:
                        Unexpected error
            security:
                - BasicAuth: []
                - BearerAuth: []
                - OAuth2:
                      - HeadlessCommerceAdminCatalog.write
            summary: Creates or updates a Sku related to a Product.
            tags:
                - Sku
...
    /skus/by-externalReferenceCode/{externalReferenceCode}:
        delete:
            operationId: deleteSkuByExternalReferenceCode
            parameters:
                - in: path
                  explode: false
                  name: externalReferenceCode
                  required: true
                  schema:
                      description:
                          The resource external reference code
                      example: AB-34098-789-N
                      minLength: 1
                      type: string
                  style: simple
            responses:
                204:
                    description:
                        No content
                401:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description:
                        Authentication information is missing or invalid
            security:
                - BasicAuth: []
                - BearerAuth: []
                - OAuth2:
                      - HeadlessCommerceAdminCatalog.write
            summary: Deletes a Sku by external reference code.
            tags:
                - Sku
        get:
            operationId: getSkuByExternalReferenceCode
            parameters:
                - in: path
                  explode: false
                  name: externalReferenceCode
                  required: true
                  schema:
                      description:
                          The resource external reference code
                      example: AB-34098-789-N
                      minLength: 1
                      type: string
                  style: simple
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Sku'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/Sku'
                    description:
                        Successful operation
                400:
                    description:
                        Invalid input
                401:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description:
                        Authentication information is missing or invalid
                404:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description:
                        The specified resource was not found
                500:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description:
                        Unexpected error
            security:
                - BasicAuth: []
                - BearerAuth: []
                - OAuth2:
                      - HeadlessCommerceAdminCatalog.read
            summary: Gets a Sku by external reference code.
            tags:
                - Sku
        patch:
            operationId: patchSkuByExternalReferenceCode
            parameters:
                - in: path
                  explode: false
                  name: externalReferenceCode
                  required: true
                  schema:
                      description:
                          The resource external reference code
                      example: AB-34098-789-N
                      minLength: 1
                      type: string
                  style: simple
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/Sku'
                    application/xml:
                        schema:
                            $ref: '#/components/schemas/Sku'
                required: true
            responses:
                202:
                    description:
                        Async
                204:
                    description:
                        Updated
                400:
                    description:
                        Invalid input
                401:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description:
                        Authentication information is missing or invalid
                404:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description:
                        The specified resource was not found
                500:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description:
                        Unexpected error
            security:
                - BasicAuth: []
                - BearerAuth: []
                - OAuth2:
                      - HeadlessCommerceAdminCatalog.write
            summary: Updates a Sku by external reference code.
            tags:
                - Sku
...
servers:
    - description:
          SwaggerHub API Auto Mocking
      url: https://virtserver.swaggerhub.com/liferayinc/commerce-admin-catalog/v1.0